MNT: Use GEMV in enet_coordinate_descent (#11507)
Make use of the BLAS GEMV operation in `enet_coordinate_descent` instead
of using DOT in a `for`-loop. Go ahead and use GEMV with both
non-transposed and transposed arrays. Previously we have had issues with
the vendored BLAS and GEMV on transposed arrays, but this attempts to
use GEMV on transposed arrays anyways. Hopefully we can make them work
as well.

As GEMV and DOT in a `for`-loop are both semantically equivalent, this
is a reasonable change to make.  Though GEMV likely uses a multithreaded
approach unlike our application of DOT in a serial loop here. In BLAS
implementations that do use threads for DOT, we can expect that GEMV
will make better usage of those threads and avoid unnecessary setup and
teardown costs that DOT in a `for`-loop is likely to incur (possibly in
each iteration of the `for`-loop).