Refactor linear assignment algorithm in cython.
The initial of this is just to transport the original python code of `linear assignment` into cython. however, i am not very familiar with cython and the code need to be optimized a lot. I sincerely hope that people from the community can help me optimize this program and further improve this algorithm.

> I haven't add compile set in setup.py as to request experts who may want to optimize this cython code, really appreciate it and thanks!