Make the `eps` parameter 'warn', as there is no good default
The old default value of 0.5 is quite arbitrary.
It may work well for 2D normalized toy examples, but for real data the user really should choose this parameter, not rely on the default to be "usually good" - it won't be good most of the time.

Hence, warn the user if he does not specify the eps parameter.