Added control structure to handle
Test stub created for when normalizer has
very small value, yet to find a test case
for h_true and h_pred which would return a
normalizer with a very small value.

On branch nmi_bug_fix
Your branch is up to date with 'origin/nmi_bug_fix'.

Changes to be committed:
	modified:   supervised.py
	modified:   tests/test_supervised.py