PERF Shrink arrays to size in liblinear helper dense_to_sparse
Calling realloc to shrink the size of a previously allocated array is
guaranteed to succeed. This way is faster because the array typically
does not need to be copied to a new location, it requires less memory
because there is no temp array that has to coexist with the last array
to be allocated, and it eliminates the need for gotos.