lambda_ in PowerTransformer.inverse_transform doc
The current docstring of `PowerTransformer.inverse_transform` contains syntax errors.
`lambda` is a reserved keyword in Python, used to express anonymous functions (as in Church-Turing "lambda calculus").
This commit replaces `lambda` by `lambda_` in order to circumvent this reserved keyword.
Note that `scipy.stats.boxcox` uses `lmbda` (without the "a"), but the trailing underscore in `lambda_` is more consistent with the attribute of sklearn's `PowerTransformer` class (`lambdas_`). It's also PEP8-compliant.