Updated the behaviour of get_feature_names to treat trans='passthrough' the same way as remainder='passthrough'.
The behaviour of get_feature_names for passthrough is now the following:
- If fitted on a dataframe, then the columns passed with the trans='passthrough'
will be treated as positional if int (in which case the feature name will be
the column name at that position) or they will be used for the actual feature name.
Any columns in remainder='passthrough' will be appended to the end of feature_names,
as the column names from the fitted dataframe
- If fitted on an array, then the column names for both when trans='passthrough' or
when remainder='passthrough' will be 'xi' for each index value i. In terms of ordering,
the remainder columns will again come after the rest of the feature names.