Bug in LinearClassifierMixin.decision_function()
This function is suppoed to calculate the signed distance from a point to various hyperplanes (i.e. the decision boundaries). However, the current formula used to calculate the score is incorrect as it does not divide the projection of the point on the hyperplane by the magnitude of the normal vector to the hyperplane. Since each hyperplane can have a different magnitude, this can affect LinearClassifierMixin.predict() and LinearClassifierMixin.predict_proba_lr() causing an incorrect prediction. This is a fairly major bug affecting all downstream classifiers (e.g. LinearSVC and LogisticRegressuib). Happy to provide an example if more information is needed.