This addresses part of #11991
The function
sklearn.datasets.base._pkl_filepath
is a compatibility function for Python 2 support.

The function expects path components as *args and constructs and returns the path to a pickel from it.

It modifies the last path component as follows:
- For Python 3, it adds a suffix ('_py3' by default, or the value of the kwarg 'py3_suffix' if specified).
- For Python 2, it does nothing .
It then concatenates its arguments using os.path.join and returns the
result.

If Python 2 support is to be deprecated, we can
- remove the function _pkl_filepath and
- replace all calls to it by direct calls to os.path.join.