Solves part of #11991
The function
sklearn.datasets.base._pkl_filepath
is a compatibility function for Python 2 and Python 3 support.

The function expects path components as *args and constructs and returns the path to a pickel from it.

It modifies the last path component as follows:

For Python 3, it adds a suffix ('_py3' by default, or the value of the kwarg 'py3_suffix' if specified).
For Python 2, it does nothing .
It then concatenates its arguments using os.path.join and returns the
result.

If Python 2 support is to be deprecated, we can simplify the code: Just
always add the suffix. We cannot get rid of the function entirely since
some users still might have Python 2 pickles on their filesystem.

We can also shorten the docstring.