Remove pytest.importorskip from examples
All calls related to matplotlib have to be skipped here, because in
general it need not be installed. This fact is also reflected by some
CI environments.

Although using 'doctest: +SKIP' instead of pytest.importorskip also
skips the call to metrics.plot_roc_curve during tests in environments
where matplotlib is in fact available, the main goal here is to have a
clean example.