Use correct n_components
Before n_components was updated something like this:
n_components = np.searchsorted(ratio_cumsum, n_components) + 1
There is a catch in numpy's searchsorted function. It returns index i where a[i-1] < v <= a[i]. So here i can be returned such that v < a[i] which is exactly what is mentioned in the documentation: 
If 0 < n_components < 1 and svd_solver == 'full', select the number of components such that the amount of variance that needs to be explained is greater than the percentage specified by n_components.
so adding 1 here is redundant. 

But if v = a[i] then 1 should be added.

Hence I feel using np.searchsorted for this task isn't correct and it should be replaced by alternatives like:
n_components = np.nonzero(ratio_cumsum > n_components)[0][0]

Now, n_components is just the exact total number of principal components required.