Change number of classes in lfw_pair documentation
Previous number of 5749 was actually the size of the set from which the pair of face labels were sampled. It was misleading because the task isn't to predict the label of the pair but wether or not the two faces belong to the same person, hence a binary classification task.