New Setup file (top level) for TEST Pypi upload
Replace the setup.py in the root level (NOT in sklearn folder).
Used to add the project source to Test Pypi.
NOTE: the binary wheel building is pointless as Test Pypi does not let us upload a simple linux version and we cannot create a many-linux wheel (which may just work). So only run the sdist and don't waste time with bdist_wheel.