Simplify definitions of types for numpy.
According to
https://github.com/scikit-learn/scikit-learn/pull/17228/commits/9e9cdb0192394d3797f4a1adc8c28db7ee05acf2#issuecomment-638998985
these dictionaries with offsets are there for old numpy versions and
they are no longer needed. We can use instead an approach similar to
https://github.com/scikit-learn/scikit-learn/pull/15097/files#diff-b071106a87a03ee4e9149e3f0a2b1180L187

I have verified that tests keep passing and that the undefined behavior
that motivated this pull request in the first place is still fixed.