DOC Avoid atypical usage of cross_validate
Training a model before passing it to cross_validate is not wrong, but also is not necessary. A user can be expected to infer typical usage or best practices from examples in the manual, so this usage should be avoided in the manual in this example.

I have verified that the results of this REPL session do not change with the modification I am proposing. I am using sklearn version 0.22.1.