RFECV Adding customizable selecte feature size method with visualization
#### Reference Issues/PRs

#### What does this implement/fix? Explain your changes.
In real world cases, where the selected feature size must be smaller than a certain number (i.e. Biomarker selection, COVID-19 PCR probe design, Sequencing panel design and etc.) or to minimize the feature size is more important than maximize the performance scores(i.e. 20 features with ROC-AUC of 0.97 and 200 features with ROC-AUC of 0.98),there is currently no such method to easily do it without going to the original code of the scikit-learn orbuild a RFECV from scratch. 

This adding feature allows mannually selection of the feature size with the visualized comparision to the optimalfeature size. Also there is no compromises to the performance if the feature is truned off.

#### Any other comments?