Use check_finite=False in _check_precision_matrix
We are safe removing the check for non-finite values because
`_check_array` is taking care of it.

To give the details:
 - there are 2 usages of `_check_precision_matrix` in
`_gaussian_mixture.py` which are preceded by a call to `_check_array` on
its arguments.
 - there is 1 usage of `_check_precision_matrix` in
`_bayesian_mixture.py` which is preceded by a call to `check_array` on
its argument.