Update _weight_boosting.py
My idea: 
Add the feature of resample (like in Bagging Classifier) to AdaBoost. In the literature this is quite common.
In every iteration of the n_estimators Boosting n_samples will be randomly picked and used for the training of the base estimator but the error score evaluation will be performed on the whole data set. 
This way overfitting is largely reduced.