Add return list order for function `train_test_split`
I found the doc about the returned list for function `train_test_split` is not clear enough. So I add the order of the returned list as

> 'X_train, X_test, y_train, y_test.'

Wish this could help beginners like me. Of course, I could know this info from the example below. But I still think it is better for a notation for the list order explicitly.