Use stable numpy scipy release for [icc-build] and [arm64] on travis
The scipy-dev builds have moved to Azure Pipelines and there is no
reason to not use stable versions of numpy and scipy to run the ICC
and ARM64 tests on travis.

This should fix the invalid wheel metadata failure observed on travis
which was already resolved on Azure Pipelines by using the legacy
pip dependency resolver for the scipy-dev build.