Fix top_k_accuracy_score for multiclass case
If a problem is actually "multiclass", and not all classes are contained
in the parameter `y_true` , the function fails, because the
determined type is "binary".
That decision makes sense, if the parameter labels is not passed.
The problem is that the function also fails, if the parameter `labels`
is passed, although it would be possible to determine the type of and
the number of classes in conjunction with this parameter.
This commit fixes that, by checking whether the `labels` parameter has
been set and contains more than 2 classes, if the type has been
determined to be "binary" in the previous step.