Adjust neighbors:test_kde_sample_weights for arccos distance (#7829)
In one-dimensional data, the assertions fail because NaN values appear
in different places. Skipping arccos for d==1 seems defensible,
because the concept of angles degenerates at one dimension.

In ten-dimensional data, the effect of the sample weights was just
under the original magic number of 0.001, at 0.0008073349365247395. As
the original threshold was an undocumented number, I presumed that it
was somewhat arbitrary, which makes lowering a bit seem defensible. I
fully realize, however, that I might be wrong about this.