sklearn: neural_network: introducing SOM method
This is a new method that can be considered a neural network but it is
commonly used as a method of clustering data. SOM is also known as
Kohonen's Self-Organizing Map. The method is based on an approximation
of a neuron's networks (weights with a topology) that are activated
according each input. For more details, see the `Reference` into the
python class description.

Signed-off-by: Julio Faracco <jcfaracco@gmail.com>