examples: neural_network: add two examples for SOM
This commit adds two examples of SOM usage: one is for color quantization of
a random image. The SOM architecture can be adjusted according color scores.
And the second one is a fork of a K-means example that can be extended to SOM
which shows some defects for data classifying as K-means does.

Signed-off-by: Julio Faracco <jcfaracco@gmail.com>