Update ensemble.rst
I don't believe that negative binomial and negative multinomial distributions are used in these loss functions. The documentation is thus potentially confusing because it seems like it is saying that a binomial deviance uses the log-liklihood of a negative binomial distribution.  So I move the word 'negative' before 'log-likelihood.' Hopefully this makes it more clear that we're just talking about treating -1 * log logliklihoods as the loss function.