[WIP] Add RadiusNeighborhood
This creates a double freeing as the buffers
are tight to vectors attached to the instance.

When the RadiusNeighborhood instance gets
deleted, so do the buffers and the content
or returned numpy array makes no sense.

Also it crashes when ones deallocates the numpy
arrays.

One needs to have the final buffers life-time
no be tight to the instance's vectors.