Do not push if the element is identical to the largest
ArgKMin context: shall there be several candidates for the k-th position
(in the k-th nn context, the furthest neighbors of the query points),
the first candidate will be picked and won't be replace the by others
candidates (which is what was happening before).

This allow matching the current behavior of k-th NN search using trees.