Inject placeholder value for MeanShift.bandwidth
This modify the test configuration so that it
makes sense for when a sole sample is provided
for MeanShift.

This test was passing previously for this
configuration but was not supposed to.

The new implementation strategy for kneighbors
which uses PairwiseDistancesArgKmin (see #21462)
is numerically stabler for this case,
motivating this modication.

Co-authored-by: Guillaume Lemaitre <g.lemaitre58@gmail.com>
Co-authored-by: Jérémie du Boisberranger <jeremiedbb@users.noreply.github.com>