Escape `name_details` and add missing =
The generated HTML generates lots of warnings in my text editor because `name_details` was not escaped. I'm not sure, but it may also be a security flaw, since someone could create a custom estimator with a `__str__()` method that generates some dangerous HTML code?

Also, there's a `=` sign missing: `class"sk-top-container"` instead of `class="sk-top-container"`.