Update _data.py
Hi,
This is my first pull request for this project. I changed part of the description because I noticed it mentioned standardization as a means to make the data look more Gaussian or like the standard normal distribution. However, standardization does not make the distribution of data any more Gaussian. It does convert the mean to 0 and standard deviation d to 1 but it does not make the distribution more symmetric.

I also changed a small grammatical error in the next paragraph.
Best,
Behrouz