MAINT Try to increase ccache hit rate on Azure Pipelines
Azure does not seem to update existing cache with the same name.
Therefore it seems that the ccache folder contents can get stalled
and the hit rate seems to diminish over time.

Base on my understanding of
https://docs.microsoft.com/en-us/azure/devops/pipelines/release/caching?view=azure-devops
it seems that if a segment of the key is a filename (and hopefully a
directory?), the contents of the file (assumed to be relative to
$(Pipeline.Workspace)) will be used to check if an updated version
of the cache blob needs to be re-uploaded.

By putting the ccache foldername as a cache key segment, I hope
that the cache will be updated whenever a compiled binary artifact
has been updated.