MAINT Improve upcast
Previously, upcast was done in the critical region.
This causes an unneeded upcast for one of the buffers.

This only upcasts buffers when necessary and where
necessary without duplication contrarily to previously.

Two methods are introduced to perform this upcast
for each strategy.

Yet, this adds some complexity to the templating.