Add sample_weight argument to _alpha_grid
Modifies _alpha_grid function in linear_model._coordinate_descent to accept a sample_weight argument.
In addition to adding this argument, I have removed the argument copy_X, which no longer seems to be needed.

The function _alpha_grid is called in two places, enet_path and LinearModelCV.
The new sample_weight argument is not used by enet_path, but it is used by LinearModelCV.