Resolves issue #14613
This commit updates the fit(X, y) definition, in the docs, of the sklearn.covariance.EllipticEnvelope class.
The original definition stated that parameter X could be of type 'array-like' or 'sparse matrix', however,
an error would be thrown if a 'sparse matrix' is passed ("TypeError: A sparse matrix was passed, but dense
data is required.").  To resolve this, 'sparse matrix' was removed from the list of types X could be, in the
docstring.