doc: micro,wtd,macro avg recall_score same as....
Clarify that the recall_score with average='micro' gives same result as with average='weighted', and both give same result as accuracy_score! It is important to know that there is no need to look at more than one of these three metrics/variants since all three are in fact identical, not distinct.

And also clarify that the recall score with average='macro' give the same result as the balanced_accuracy_score !   Again important to know which seemingly-distinct metrics are in fact identical and so need not be compared separately which would be a redundant waste of the user's time.

Here's a (non-toy) multi-class example showing that for both metrics, they match in all 15 decimal places in the existing classification_report :

```
classification_report(y_tst, y_pred_tst, digits=15) =
                 precision         recall            f1-score               support

              0  0.817246835443038 0.683879510095995 0.744638673634889      3021
              1  0.829678021465236 0.696708463949843 0.757401490947817      2552
              2  0.770103092783505 0.630912162162162 0.693593314763231      1184
              3  0.331294597349643 0.844155844155844 0.475841874084919       385
              4  0.394505494505494 0.920512820512820 0.552307692307692       390

       accuracy                                      0.700345193839618      7532
      macro avg  0.628565608309383 0.755233760175333 0.644756609147710      7532
   weighted avg  0.767319254559894 0.700345193839618 0.717240526308044      7532

balanced_accuracy_score(
              y_tst, y_pred_tst) = 0.755233760175333
```

(not sure if above will display properly monospaced)

(I've inserted some spaces above to fix the misalignment of the classification_report output for this many digits. Would you like to see the confusion matrix that gives rise to the classification report above?)

These matches are of course not a coincidence. If you compare the definition of accuracy to that of micro average recall, you will see that they are one and the same, and the weighted average recall is known to be equal to the micro average recall and therefore equal to the accuracy.

Similarly the definition of macro average recall is the same as the definition of balanced accuracy used (by default) by metrics.balanced_accuracy_score: basically the simple average over classes of the proportion (of sample weight) classified correctly among the actual instances of each given class.