MAINT Introduce interfaces for PairwiseDistancesReductions
Those interfaces are meant to be used in the Python code, decoupling the
actual implementation from the Python code. This allows changing all the
private implementation while maintaining a contract for the Python
callers.

Each interface extending the base `PairwiseDistancesReduction` interface
must implement the :meth:`compute` classmethod.

Under the hood, such a function must only define the logic to dispatch
at runtime to the correct dtype-specialized `PairwiseDistancesReduction`
implementation based on the dtype of X and of Y.

This refactoring will ease other dtype support such as float32 support.