Updating documentation of fetch functions
Updated documentation of fetch_lfw_people and fetch_lfw_pairs to include the target_names array, which was mentioned in the user guide (https://scikit-learn.org/stable/datasets/real_world.html#labeled-faces-in-the-wild-dataset), but missing in the documentation (https://scikit-learn.org/stable/modules/generated/sklearn.datasets.fetch_lfw_people.html#sklearn.datasets.fetch_lfw_people). 

Personally, since target_names for fetch_lfw_pairs is has only 2 values I think it's more intuitive to omit the target_names entirely for fetch_lfw_pairs and explain in the documentation of the target array that 0 corresponds to different persons and 1 corresponds to same person, but there might be consistency reasons to keep it that way, so I did not want to make any changes to that.