Fix conversion of `FiniteStatus` from C to Python
Currently when Cython converts a `cpdef enum` from C to Python, it
constructs a Python `int` instead of using the `IntEnum`-based
`FiniteStatus` Python object created by `cpdef`. As `IntEnum` values can
be compared with `int`, this doesn't cause issues atm. That said, to
avoid issues it would be good to return `FiniteStatus` objects to Python
code. This adds some code to manually perform this conversion between C
and Python code to workaround this Cython issue.