[WIP] Make random_state accept np.random.Generator
Done:

- Added tests for estimators
- Made tests for estimators pass

Missing

- splitters
- other components, e.g. for creating random datasets
- documentation
- docstrings

Caveats

It's almost impossible to have a complete test coverage for this
feature. The reason is that even though we check all estimators that
support random_state, we don't know if the code path that actually uses
random_state is being taken or not, since it might depend on
hyper-parameters.