MAINT Test second alternative for sparse-dense support
The modulo trick is costly because it uses integer division
under the hood which are up to 80 times slower.

An alternative is to pass the address of the `indices` array and shift
this address at the callers level with the information that we have
there so that the dereferencing works properly.

This commit replaces the modulo trick by this alternative and adapt
documentation and comments in this regard.

See:
https://github.com/scikit-learn/scikit-learn/pull/23585#issuecomment-1190672626