Revert changes related involving rng_integers
We decided to opt in estimators (and all the rest) step by step into
using Generators. Therefore, I reverted all the changes in the actual
estimators that were necessary to accomodate Generators, which comes
down to the use of the rng_integers for now.

The common test has been adjusted to have a long list of excluded
estimators -- currently containing all estimators -- that are skipped
for testing. The idea is that if a new PR comes along that opts an
estimator in, it should be as easy as crossing that estimator off the
list to be able to check if it still works.

Note for future developers. The "random_state" variable is sometimes
also referred to "rng" or "rnd" (and maybe others that I missed), so a
simple grep for "random_state" is not enough.