~50% improvement!
Downsides: we are currently using auxilary vectors declared in BestSplitter and we have to hold two whole matrices X_ranks and X_order. Further strategies to reduce memory usage are welcome.

this branch:
n_samples=1000 with 0.066 +/- 0.017
n_samples=5000 with 0.491 +/- 0.124
n_samples=10000 with 1.276 +/- 0.048
n_samples=20000 with 3.520 +/- 0.416
n_samples=50000 with 12.558 +/- 2.260

sklearn/main:
n_samples=1000 with 0.100 +/- 0.015
n_samples=5000 with 0.972 +/- 0.279
n_samples=10000 with 2.492 +/- 0.188
n_samples=20000 with 7.367 +/- 1.070
n_samples=50000 with 26.034 +/- 3.585