Bugfix: other fit_params when sw not supported
There was a bug that fit_params where not passed to fit of the base
estimator if the base estimator does not support sample_weight.

It is a bit difficult to test this. I decided to go with a test that
uses a subclassed CheckingClassifier that does not support
sample_weight (i.e. does not have it in the signature of fit). This way,
we can test the code branch that resulted in the bug.