Added note about multiple transformers on a column
As it is not clear in the current documentation that each column can only be addressed by exactly one transformer in one instance of ColumnTransformer, I propose to add a note like this:
It is not possible to address the same columns with more than one
transformer. However, multiple ColumnTransformer objects can be chained.