Implement changes as discussed with Adrin
Reworking when sample_weight is being used in BaggingClassifier and
BaggingRegressor.

They provide two mechanisms to achieve sampling of rows, first via
sample_weight (the preferred mechanism) and second via indexing (i.e.
selecting the rows from the data).

The new metadata routing information is now being used to determine of
the base estimator can actually deal with sample_weight (the assumption
is made that it does the correct thing). This is a change to the current
situation, where the signature is checked for the presence of the
"sample_weight" argument instead. With this change, in the future it
will be possible to support the sample_weight mechanism when using
metaestimators even if they don't have sample_weight in their signature.