Only compute bounds' middle_value when needed
Also simplify the logic and its comment.

Note: we could factor this duplicated logic in a private function
or a private method, but we would need to use pointers to modify
variable on the program stack because we can't simply return
several values. This factorization might be more complex than
the current duplication, hence we propose leaving it as is.

See:
https://github.com/scikit-learn/scikit-learn/pull/13649/files#r973633867

Co-authored-by: Olivier Grisel <olivier.grisel@ensta.org>