Switch to unified engine API, add multi engine support
This renames the methods an engine has to/can implement along the
pattern of `pre_X`, `X`, `post_X` (X=fit, transform, ..).

Changes the engine loading mechanism to return all engines in a user
defined order so that we can try ecah engine in turn to see which one
supports the given input case.