FIX Conditionally resolve metric deprecated by SciPy
SciPy might deprecated some distance metrics overtime,
so with need to have condtional fallback for deprecated
metrics. This is currently the case for `"kulczynski1"`.

This adds a fallback on `DistanceMetric{,32}`.

Note that for now, the newly introduced
`DistanceMetric{,32}` fallback has precedence over the
`scipy.spatial.distance` fallback, but this must not be
the case.

One way to remove this precedence would be to check
for metric inclusion in a list equals to
`scipy.spatial.distance._METRICS_NAMES` (which can't
be imported unfortunately :( ).

Hence the proposed `_SCIPY_DEPRECATED_METRICS`.