Gaussian Process: change default behavior of `sample_y`
The `random_state` argument was set to `0` forceing `sample_y` to always
return the same gaussian process samples.
This is contrary to the expected behavior most people have from working
with random number generating libraries such as numpy and scipy.
Changing the default parameter to `None` will make the behavior of
`sample_y` conform to the usual expectation a user.

Updated `test_gpr.py` so that all calls to `sample_y` are passed
`random_state=0` to enusre deterministic testing.